Move static into internal
