+++ internal/api/images.go
@@ -6,9 +6,9 @@ import (
 	"os"
 	"strings"
 
+	"github.com/stashapp/stash/internal/static"
 	"github.com/stashapp/stash/pkg/logger"
 	"github.com/stashapp/stash/pkg/manager/config"
-	"github.com/stashapp/stash/pkg/static"
 	"github.com/stashapp/stash/pkg/utils"
 )
