+++ internal/static/embed.go
